LayoutTests/imported/w3c/web-platform-tests/encoding/textdecoder-copy.any.worker-expected.txt


FAIL Modify buffer after passing it in (ArrayBuffer) assert_equals: expected "@" but got "﻿@"
FAIL Modify buffer after passing it in (SharedArrayBuffer) assert_equals: expected "@" but got "﻿@"
